update 2024 sol 8
as noted in https://redd.it/1h9boli (and a few other posts), it is
possible that the antinodes could be in between the antennas. While my
input (and seemingly other inputs as well) does not have such case, for
completeness the related code is added. This slows down the whole script
by about 0.3ms.